Move leaders to the leaders directory
