+++ Vic2ToHoI4/Source/V2World/Country.cpp
@@ -4,7 +4,7 @@
 #include "Localisations/Vic2Localisations.h"
 #include "Log.h"
 #include "Military/ArmyFactory.h"
-#include "Military/Leader.h"
+#include "Military/Leaders/Leader.h"
 #include "ParserHelpers.h"
 #include "Politics/Party.h"
 #include "Pops/Pop.h"
