+++ Vic2ToHoI4/Source/V2World/Country.h
@@ -10,8 +10,8 @@
 #include "Diplomacy/RelationsFactory.h"
 #include "Localisations/Vic2Localisations.h"
 #include "Military/Army.h"
-#include "Military/Leader.h"
-#include "Military/LeaderFactory.h"
+#include "Military/Leaders/Leader.h"
+#include "Military/Leaders/LeaderFactory.h"
 #include "Parser.h"
 #include "Politics/Party.h"
 #include "Provinces/Province.h"
