+++ Vic2ToHoI4/Source/V2World/Military/Leaders/LeaderFactory.cpp
